Remove dependency of type system tests on ILCompiler.ReadyToRun
This is now shared unit test project that is not AOT/JIT specific.

The recently added MarshalUtilsTests.cs created a dependency on these.
