+++ src/coreclr/tools/Common/TypeSystem/Interop/IL/Marshaller.cs
@@ -11,63 +11,10 @@ using ILLocalVariable = Internal.IL.Stubs.ILLocalVariable;
 
 namespace Internal.TypeSystem.Interop
 {
-    enum MarshallerKind
-    {
-        Unknown,
-        BlittableValue,
-        Array,
-        BlittableArray,
-        Bool,   // 4 byte bool
-        CBool,  // 1 byte bool
-        VariantBool,  // Variant bool
-        Enum,
-        AnsiChar,  // Marshal char (Unicode 16bits) for byte (Ansi 8bits)
-        UnicodeChar,
-        AnsiCharArray,
-        ByValArray,
-        ByValAnsiCharArray, // Particular case of ByValArray because the conversion between wide Char and Byte need special treatment.
-        AnsiString,
-        UnicodeString,
-        UTF8String,
-        AnsiBSTRString,
-        BSTRString,
-        ByValAnsiString,
-        ByValUnicodeString,
-        AnsiStringBuilder,
-        UnicodeStringBuilder,
-        FunctionPointer,
-        SafeHandle,
-        CriticalHandle,
-        HandleRef,
-        VoidReturn,
-        Variant,
-        Object,
-        OleDateTime,
-        Decimal,
-        OleCurrency,
-        Guid,
-        Struct,
-        BlittableStruct,
-        BlittableStructPtr,   // Additional indirection on top of blittable struct. Used by MarshalAs(LpStruct)
-        LayoutClass,
-        LayoutClassPtr,
-        AsAnyA,
-        AsAnyW,
-        ComInterface,
-        BlittableValueClassWithCopyCtor,
-        Invalid
-    }
     public enum MarshalDirection
     {
         Forward,    // safe-to-unsafe / managed-to-native
         Reverse,    // unsafe-to-safe / native-to-managed
-    }
-
-    public enum MarshallerType
-    {
-        Argument,
-        Element,
-        Field
     }
 
     // Each type of marshaller knows how to generate the marshalling code for the argument it marshals.
